Refactor LLM client architecture
