+++ docs/architecture/integrations.md
@@ -16,6 +16,13 @@
 - Без ключа или при отключённых реальных тестах (`COOKAREQ_RUN_REAL_LLM_TESTS` не установлен) сценарии в `tests/integration/test_llm_openrouter_integration.py` пропускаются, чтобы не генерировать ошибочные обращения.
 - Ошибки валидации инструментов мапятся через `app.mcp.utils.exception_to_mcp_error` в финальный ответ агента, что предотвращает бесконечные повторы при некорректных tool calls.
 
+### LLM client
+- `LLMClient` остаётся тонким фасадом, который выбирает стратегию (chat/harmony/stream) и делегирует подготовку запросов и разбор ответов специализированным компонентам.
+- `LLMRequestBuilder` отвечает за подготовку payload: нормализует историю, режет её по лимиту токенов, собирает системные блоки и преобразует формат сообщений (например, в Qwen-сегменты). Отдельный метод строит Harmony prompt на основе того же пайплайна.
+- `LLMResponseParser` принимает сырой ответ от бэкенда, нормализует tool calls, восстанавливает JSON аргументов, собирает reasoning-сегменты и конвертирует их в `LLMReasoningSegment`/`LLMToolCall`. Логика восстановления и телеметрия по аргументам теперь сосредоточены здесь.
+- Модуль `app.llm.logging` предоставляет `log_request`/`log_response`, чтобы фасад просто прокидывал полезную нагрузку, а детали телеметрии хранились в одном месте.
+- Для юнит-тестов добавлены заглушки `tests/unit/llm/factories.py`: они позволяют подменить `LLMRequestBuilder`/`LLMResponseParser` и собрать синтетический `LLMResponse`, когда важно протестировать потребителей LLM без реальных сетевых обращений.
+
 ## 2. MCP-сервер и инструменты
 - [ ] Архитектура MCP-сервера (FastAPI, IPC, запуск из GUI/CLI).
 - [ ] Таблица зарегистрированных инструментов и их контрактов.
