+++ tests/unit/llm/__init__.py
@@ -1,0 +1,1 @@
+"""Helper stubs for unit tests interacting with LLM components."""
